app/services/message_processor.py
import logging
from typing import Optional, Dict, Any
from datetime import datetime

from aiogram import types
from tortoise.transactions import in_transaction

from app.models.db_models import User, Chat, Message, Media, ChunkEmbedding

logger = logging.getLogger(__name__)

async def process_message(message: types.Message) -> Optional[str]:
    """Process an incoming message and return a response if needed"""
    return None

async def save_message(message: types.Message) -> Message:
    """Save a message to the database"""
    return None

async def create_chunks_for_chat(chat_id: int):
    """Create chunks for messages in a chat"""
    return None
