requirements.txt
aiogram
aiohttp
python-dotenv
redis

tortoise-orm
tortoise-orm[asyncpg]
asyncpg
aiosqlite
tortoise-vector==0.1.4
pgvector

openai
langchain
langchain-openai
numpy

# Development
aioresponses
pytest
pytest-asyncio
pytest-cov
black
isort
mypy
flake8
pytest-mock

gunicorn
